+++ src/triangulation/mod.rs
@@ -1,0 +1,3 @@
+mod delaunay2;
+
+pub use delaunay2::{Triangulation2, delaunay2};